Add output.json schema handling
This schema can be used as a template for the output structure.
In the case of this FHIR Patient being created for bahmni the FHIR
resource type needs to be included but is not present in the input from
the user. Therefore it is included in the output object template.

OHM-871