Makes certain files in an endpoint folder optional for startup
Prior to this change, all endpoint files were required for the
mediator to start up: meta, input-validation, input-mapping and
constants.

This change attaches a mandatory/optional flag to be attached to
each of the four files. If the file is specified as optional and
it does not exist in the endpoint folder, the mediator will
start as normal.
If a required file is not present, the mediator will (still)
abort with an error message.

OHM-893