Fix .gitignore file
The git ignore file excluding subdirectories of the endpoints directory
caused tests to fail.

We don't want the subdirectories of the endpoints directory stored to
github as these directories are user specific. Unfortunately, this
empty directory has to exist. Users need it to put their config files
in. The tests need it as they create config files during the integration
tests.

Without content the endpoints directory does not exist on github.
Therefore I've added the config.md file to the subdirectory to ensure it
exists. This file contains the config structure needed for the directory

OHIE-296