Add check within routes to skip files in directory
Only subdirectories are used to create routes.
Files immediately with the endpoints directory can be ignored.
This is necessary because there is a config.md file included in the
directory. This file is only there to ensure the directory has content.
Without content the directory is not registered in github. This causes
errors with travis as well as docker :(

OHIE-296