fix the breaking tests
The functions under test have been modified. The orchestration response and request bodies have been changed to string from objects so the expected results in the tests had to be modified. The request header value 'x-openhim-transaction' has been changed to be lowercase as it comes in as lowercase from the openhim