fix the setting of the koa response
The response object that was being returned is not whats expected. If there is a primary or only one response request, the response data should not be inside a nested object. If there are multiple requests and none is specified as the primary, only the last response was being returned. These issues are fixed in this commit. The tests have also been modified as they were now failing due to the changes