modify method for setting the status text
Instead of having multiple variables on the context for determining the status text, we now use one variable - an array, which will contain the strings that will be used to determine the status text. The method that handles errors has been modified, differnt error types are pushed to the array. Tests have also been changed.