modify method for adding query parameters
The method for adding query parameters has been modified to extract parameters directly from the payload and the request query object. The user no longer has to have a params object in the payload. some other methods have also been modified so that the query parameters are added to the orchestrations response and request objects

TRACE-112