Add enum validation to endpoint schema
The transform fields input and output are enum string values. These
values were not being validated before being added to the DB.

This would have resulted in errors later in the code lifecycle and would
be far harder to remedy.
Failing validation at this point is far more efficient

OHIE-376