Allow empty mapper schema object
The validation and external requests middleware both cater for empty
objects, whilst the mapper would throw an error.

This behaviour is not correct as teh mapping mediator can be used for
only validation or only mapping. Therefore, users do not need to think
about mapping schemas on their endpoints if it won't be used.

However, when no schema is available we will log a warn message that
their is no mapping schema.

OHIE-374