Add Change Listener that populates endpoint cache
The Change listener is a Mongoose Model.watch function that monitors the
Db for any changes to the endpoints collection. These changes feedback
to our listener funciton which triggers a re-read of the existing
endpoints and populates this into the Endpoint Cache. (which is jsut an
in memory array - so super fast)
We shouldn't have to worry about how much memory the cache uses as our
config is made up of small json objects... If they do ever get too big
we can look at implementing a cache module dependency.

OHIE-374