Rename test files to remove 'Test' suffix
These tests are all contained within the test subdirectory therefore the
suffix 'Test' was unneccessary.
The suffix also wasn't needed to run the test from package.json as the
utils file is in the root dir whilst all the test files are in
subdirectories which allows them to be specified.
We shouldn't need any non test js files within the subdirectories as it
is good test practice to include all the testing data in the test that
needs it so that testers can see exactly what is in their test and
there is minimal change of data contamination between tests.

OHIE-375