Replace private func export with Rewire module
The validator middleware exported its private function when the NODE_ENV
was test so that the function could be tested - that seems a bit dodge -
therefore we are replacing that testing mechanism with the node module
Rewire.
This module is only active in the test files and gives access to the
private functions without any code change or logic.

OHM-994