Add timeout between test steps :( :( :(
Needed to simulate a realisic use case.
The normal speed of the running tests causes a cache race condition that
can fail tests intermittantly...

This timeout is the easiest fix. A better fix would be to make use of
the collection change listeners - however, their current setup is not
very compatible with the tests and causes even bigger test failures :(

So the timeout is the lesser evil for now.

OHM-997