correct response status and message on mapping error
The object mapping fails when the mapping schema defined is invalid (for example if it contains a function that is not in the mapper module). In this case the correct status to respond with is '500'. The error message has also been made more verbose. Tests have also been modified as result of this
OHM-998