modify tests for parser and external requests
These tests had to be modified because the parser and external requests functions have been modified. The logic for constucting the response in openhim format has been moved to the parser middleware, which is the last middleware befor the response is sent. Orchestrations are done in the parser middleware and they were not being added to the response body

OHM-998