make use of rawbody for the parsing
The library that was being used for parsing has an issue where one promise is not resolved when the request body is not supplied. One way to get around the issue is having logic that checks if there is a body being sent using the content-length, however this is not very good because the content-length header of a request can be manipulated by a user. The rawbody library is now used for parsing. It handles cases where the request body is not supplied

OHM-1013