change the database to use for endpoint tests
There are tests in the module that are nested in other tests. With the way the test servers are set up and used, the enpoints which are supposed to be deleted when the test suite tears down were not being deleted. This was causing the tests to fail after running successfully the fist time on an empty db as duplicates are not allowed. The tests have been changed to use the 'test' database which will be emptied after the tests run

OHM-1014