Remove unintended file
This file was uploaded by accident and should be removed. The steps for pushing to docker is described in a readme as automating this process with a script can have undesired consequences. The user will be very careful that they are on the correct codebases when deploying.

Safest option is to follow the simple manual steps for now