add logic to ensure the correct response is sent back
This is to ensure that when a request does not match any of the endpointz in the database, the correct response is sent back and that some actions which should only be performed when the request matches an endpoint are not executed. Some tests have also been modified as a result of the logic added.

OHIE-407