remove logic for parsing only when the endpoint exists
This logic was added because when endpoint does not exist there some errors that were happening. The errors were happening because when an enpoint is not found we were setting the response type to json but were sending back a string. The logic for checking endpoint existance was added to counter the failures and is now being removed as the source of the failure has been traced

OHIE-407