add logic for validating endpoint patterns
Url parameters are now supported and the logic that processes the retrieval of the url parameters (for use in external requests) expects the endpoint pattern to be valid, that is it should have parameter names. patterns like "/path/:/" or "/path/:" will result in the functionality that handles the url path matching failing. The logic that is added in this commit ensures that the url patterns are valid

TRACE-178