Enforce consistent koa ctx headers reference
Using the Koa context, ctx.header, ctx.request.header, ctx.headers, and
ctx.request.headers all appear to refer to the same field.

This is very confusing.

In our mapper schemas we define the headers field as `config.headers`.
To ensure project consistency all the references to to header are now
changed to headers.

Hopefully this reduces confusion and prevents issues relating to
unintentionally assigning headers to the wrong places

TRACE-180