Adds the ability send lookup requests forEach item of an array
This allows more complex workflows where we need to trigger requests on an item by item basis. This change also supports setting the concurrency for these requests so the user may tune the performance. This is the basic first commit of this work, expect more commits with better error handling and bug fixes.

TRACE-202