Remove ctx body before doing response request
This is done in the performResponseRequest function. This allows the
response to be stored in its place.
The ctx body needs to be removed separately for the array reqests as the
ctx in the array requests is a shallow copy therefore the data in that
field would never get cleared.

TRACE-207