Fix state tests to cater for bug fix
The state bug was that the lookup request state extract was not
behaving like the other data fields. The function used to extract to
data from the supplied path used Object key check on the data values
which returned the keys of fields that did not have an extract config
associated with them therefore the function was returning empty
strings for fields that shouldn't have been captured.

OHIE-451