Strip out unnecessary orchestration data from lookup data
When a mediator responds to a call that has a header indicating it
originated from the OpenHIM it will nest and stringify the response data
within the `response.body`. This is not helpful when we are chaining
requests in the mapping mediator as the data returned to us needs to be
manipulated - this is not possible if that data has been stringified.

Therefore the logic here will check whether the lookup data came from a
mediator - if so it will attempt to JSON parse the response body and
place that data into the allData.lookupRequest field. This will make it
easy to manipulated returned mediator data without affecting the real
orchestration information needed by the OpenHIM.

DSC19-25